Move logic from RemainingTiles to selectors
