+++ packages/scrabble-solver/src/components/RemainingTiles/Character.module.scss
@@ -1,6 +1,7 @@
 .character {
   display: flex;
   flex-direction: column;
+  align-items: center;
   transition: var(--transition);
 
   &.finished {
