+++ packages/scrabble-solver/src/components/RemainingTiles/RemainingTiles.tsx
@@ -1,8 +1,6 @@
-import { BLANK, CONSONANTS, VOWELS } from '@scrabble-solver/constants';
 import React, { FunctionComponent } from 'react';
 
-import { selectRemainingTiles, selectRemainingTilesCount, useTranslate, useTypedSelector } from 'state';
-import { RemainingTile, RemainingTilesGroup } from 'types';
+import { selectRemainingTilesCount, selectRemainingTilesGroups, useTranslate, useTypedSelector } from 'state';
 
 import Sidebar from '../Sidebar';
 
@@ -15,48 +13,16 @@ interface Props {
   onClose: () => void;
 }
 
-const getRemainingCount = (remainingTiles: RemainingTile[]): number => {
-  return remainingTiles.reduce((sum, { count, usedCount }) => sum + count - usedCount, 0);
-};
-
-const getTotalCount = (remainingTiles: RemainingTile[]): number => {
-  return remainingTiles.reduce((sum, { count }) => sum + count, 0);
-};
-
 const RemainingTiles: FunctionComponent<Props> = ({ className, isOpen, onClose }) => {
   const translate = useTranslate();
-  const remainingTiles = useTypedSelector(selectRemainingTiles);
-  const totalRemainingCount = useTypedSelector(selectRemainingTilesCount);
-  const consonants = remainingTiles.filter(({ character }) => CONSONANTS.includes(character));
-  const vowels = remainingTiles.filter(({ character }) => VOWELS.includes(character));
-  const blanks = remainingTiles.filter(({ character }) => character === BLANK);
-
-  const groups: RemainingTilesGroup[] = [
-    {
-      remainingCount: getRemainingCount(vowels),
-      tiles: vowels,
-      translationKey: 'results.header.vowels',
-      totalCount: getTotalCount(vowels),
-    },
-    {
-      remainingCount: getRemainingCount(consonants),
-      tiles: consonants,
-      translationKey: 'results.header.consonants',
-      totalCount: getTotalCount(consonants),
-    },
-    {
-      remainingCount: getRemainingCount(blanks),
-      tiles: blanks,
-      translationKey: 'results.header.blanks',
-      totalCount: getTotalCount(blanks),
-    },
-  ];
+  const remainingTilesCount = useTypedSelector(selectRemainingTilesCount);
+  const groups = useTypedSelector(selectRemainingTilesGroups);
 
   return (
     <Sidebar
       className={className}
       isOpen={isOpen}
-      title={`${translate('remaining-tiles')} (${totalRemainingCount})`}
+      title={`${translate('remaining-tiles')} (${remainingTilesCount})`}
       onClose={onClose}
     >
       {groups.map(({ remainingCount, tiles, translationKey, totalCount }) => (
